+++ doc/source/doxygen-docs/changelog.md
@@ -6,6 +6,8 @@
 - Changes in applications:
   - RawlogViewer:
     - Browse dialog: Smarter coloring of pointclouds; check all & none buttons for sensor layers.
+  - rawlog-edit:
+    - Operation `--camera-params` now also works for mrpt::obs::CObservation3DRangeScan observations.
 - Changes in libraries:
   - \ref mrpt_poses_grp
     - mrpt::poses::CPose3DQuat: Remove use of obsolete base class std::iterator.